Adds Samples solution, moves WpfDemo to samples, moves Covox (engine) to Covox.
