+++ samples/Samples.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31019.35
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WpfDemo", "WpfDemo\WpfDemo.csproj", "{E142D4DB-088E-4A4B-8FD7-17AE2C1357D8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E142D4DB-088E-4A4B-8FD7-17AE2C1357D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E142D4DB-088E-4A4B-8FD7-17AE2C1357D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E142D4DB-088E-4A4B-8FD7-17AE2C1357D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E142D4DB-088E-4A4B-8FD7-17AE2C1357D8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8241235B-28A8-4FB4-978C-2671F67B4C6D}
+	EndGlobalSection
+EndGlobal
