+++ samples/WpfDemo/App.xaml
