+++ samples/WpfDemo/App.xaml.cs
