+++ samples/WpfDemo/AssemblyInfo.cs
