+++ samples/WpfDemo/MainWindow.xaml
