+++ samples/WpfDemo/assets/light_off.jpg
