+++ samples/WpfDemo/assets/light_on.jpg
