+++ samples/WpfDemo/assets/window_close.jpg
