+++ samples/WpfDemo/assets/window_open.jpg
