+++ src/CoVoX.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31005.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Covox", "Covox\Covox.csproj", "{E9628E09-C233-4D15-B8BC-9577487133AF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E9628E09-C233-4D15-B8BC-9577487133AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E9628E09-C233-4D15-B8BC-9577487133AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E9628E09-C233-4D15-B8BC-9577487133AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E9628E09-C233-4D15-B8BC-9577487133AF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F163D03B-109B-4D79-9D24-92F9A5F62063}
+	EndGlobalSection
+EndGlobal
