--- src/CoVoX/CoVoX.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31005.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Covox", "API\Covox.csproj", "{E9628E09-C233-4D15-B8BC-9577487133AF}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo", "Demo\Demo.csproj", "{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WpfDemo", "WpfDemo\WpfDemo.csproj", "{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E9628E09-C233-4D15-B8BC-9577487133AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E9628E09-C233-4D15-B8BC-9577487133AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E9628E09-C233-4D15-B8BC-9577487133AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E9628E09-C233-4D15-B8BC-9577487133AF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{87F3BA2B-E5B2-474C-8D30-5A181E766CF8}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D1AF4ADA-A59D-4D08-86C4-47D046DA792E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F163D03B-109B-4D79-9D24-92F9A5F62063}
-	EndGlobalSection
-EndGlobal
