--- src/CoVoX/Demo/appsettings.json
@@ -1,34 +1,0 @@
-{
-  "Serilog": {
-    "MinimumLevel": "Debug",
-    "WriteTo": [
-      {
-        "Name": "Async",
-        "Args": {
-          "configure": [
-            {
-              "Name": "Console",
-              "Args": {
-                "theme":
-                  "Serilog.Sinks.SystemConsole.Themes.AnsiConsoleTheme::Code, Serilog.Sinks.Console",
-                "outputTemplate":
-                  "{Timestamp:yyyy-MM-dd HH:mm:ss.fff zzz} [{Level:u3}] {Message:j}{NewLine}"
-              }
-            },
-            {
-              "Name": "File",
-              "Args": {
-                "restrictedToMinimumLevel": "Warning",
-                "path": "Logs\\log.txt",
-                "rollingInterval": "Day",
-                "fileSizeLimitBytes": 10240,
-                "rollOnFileSizeLimit": true,
-                "retainedFileCountLimit": 30
-              }
-            }
-          ]
-        }
-      }
-    ]
-  }
-}
