--- src/CoVoX/WpfDemo/dummysecrets.json
@@ -1,6 +1,0 @@
-{
-  "AzureConfiguration": {
-    "SubscriptionKey": "",
-    "Region": ""
-  }
-}
