+++ src/Covox/Models/Command.cs
