+++ src/Covox/Modules/Translating/IMultiLanguageTranslatingModule.cs
