+++ src/Covox/Modules/Translating/ITranslatingModule.cs
