+++ src/Covox/Modules/Translating/MultiLanguageTranslator.cs
