+++ src/Covox/Modules/Translating/Translator.cs
