+++ src/Covox/Modules/Understanding/Interpreters/IInterpreter.cs
