+++ src/Covox/Modules/Understanding/Interpreters/TokenSimilarityInterpreter.cs
