+++ src/Covox/_Contracts/IExposeErrors.cs
