+++ src/Covox/_Internals/DefaultLogger.cs
